+++ src/main/java/de/cismet/lagisEE/entity/extension/baum/BaumKategorie.java
@@ -31,14 +31,14 @@ public interface BaumKategorie extends Serializable {
      *
      * @return  DOCUMENT ME!
      */
-    Long getId();
+    Integer getId();
 
     /**
      * DOCUMENT ME!
      *
      * @param  id  DOCUMENT ME!
      */
-    void setId(final Long id);
+    void setId(final Integer id);
 
     /**
      * DOCUMENT ME!
